Jwt_roles

+++ springbootproject/src/main/java/com/springbootproject/repository/LoggerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 
 import com.springbootproject.entity.LoggerEntity;
 @Repository
-public interface LoggerRepository  extends JpaRepository<LoggerEntity, Long>{
+public interface LoggerRepository  extends JpaRepository<LoggerEntity, Integer>{
 	
 	void  removeByToken(String token);
